model updates

+++ models/KNN/Item_KNN_CFRecommender.py
@@ -38,11 +38,11 @@ class ItemKNNCFRecommender(RecommenderSystem, RecommenderSystem_SM):
                 similarity = Compute_Similarity(self.URM_train, **similarity_args)
             else:
                 if self.feature_weighting == "BM25":
-                    self.URM_train_copy = self.URM_train.astype(np.float64)
+                    self.URM_train_copy = self.URM_train.astype(np.float32)
                     self.URM_train_copy = to_okapi(self.URM_train)
 
                 elif self.feature_weighting == "TF-IDF":
-                    self.URM_train_copy = self.URM_train.astype(np.float64)
+                    self.URM_train_copy = self.URM_train.astype(np.float32)
                     self.URM_train_copy = to_tfidf(self.URM_train)
                 similarity = Compute_Similarity(self.URM_train_copy, **similarity_args)
         else:
@@ -50,11 +50,11 @@ class ItemKNNCFRecommender(RecommenderSystem, RecommenderSystem_SM):
             self.shrink = shrink
             self.feature_weighting = feature_weighting
             if self.feature_weighting == "BM25":
-                self.URM_train_copy = self.URM_train.astype(np.float64)
+                self.URM_train_copy = self.URM_train.astype(np.float32)
                 self.URM_train_copy = to_okapi(self.URM_train)
 
             elif self.feature_weighting == "TF-IDF":
-                self.URM_train_copy = self.URM_train.astype(np.float64)
+                self.URM_train_copy = self.URM_train.astype(np.float32)
                 self.URM_train_copy = to_tfidf(self.URM_train)
             if self.feature_weighting == "none":
                 similarity = Compute_Similarity(self.URM_train, shrink=shrink, topK=topK, normalize=normalize,
